Phase 6.4: Extract useTaskMeasurement hook from useTaskDerived
Moves taskWidths state, taskElementRefs ref, ResizeObserver effect, and
setTaskRef callback into a dedicated useTaskMeasurement hook. useTaskDerived
now imports and delegates to it.
